Refactor - Move refactor.utils function to the utils package
This permit to have less cross dependency as the utils package
does not depend on anything. The checker package still depends
on reporter. Also moved classes from __init__ to their own file
in reporter.
